+++ aquadoggo/src/graphql/convert/query.rs
@@ -2,47 +2,32 @@
 
 use std::convert::TryInto;
 
-use anyhow::Result;
 use apollo_parser::ast::{
     Argument as AstArgument, AstChildren, Definition, Document, Selection, Value,
 };
-use apollo_parser::Parser;
-use p2panda_rs::hash::{Hash, HASH_SIZE};
-use sea_query::{Alias, Expr, Query};
+use p2panda_rs::hash::Hash;
 use thiserror::Error;
 
-const MAX_SCHEMA_NAME: usize = 64;
-const SCHEMA_NAME_SEPARATOR: &str = "_";
-
-fn parse_graphql_query(query: &str) -> Result<Document> {
-    let parser = Parser::new(query);
-    let ast = parser.parse();
-
-    if ast.errors().len() > 0 {
-        panic!("Parsing failed");
-    }
-
-    Ok(ast.document())
-}
+use crate::graphql::convert::{Schema, SchemaParseError};
 
 type MetaFields = Option<Vec<MetaField>>;
 type Fields = Option<Vec<Field>>;
 type Arguments = Option<Vec<Argument>>;
 
 #[derive(Debug)]
-enum Argument {
+pub enum Argument {
     /// Filter document views by document hash.
     DocumentHash(Hash),
 }
 
 #[derive(Debug)]
-enum MetaField {
+pub enum MetaField {
     /// Hash of each document view.
     DocumentHash,
 }
 
 #[derive(Debug)]
-enum Field {
+pub enum Field {
     /// Name of the application data field to query.
     Name(String),
 
@@ -51,87 +36,22 @@ enum Field {
 }
 
 #[derive(Debug)]
-struct Relation {
+pub struct Relation {
     /// Schema hash of the related document view.
-    schema: Schema,
+    pub schema: Schema,
 
     /// Fields which give general information on the related document view.
-    meta_fields: MetaFields,
+    pub meta_fields: MetaFields,
 
     /// Materialized field values of the related document view.
-    fields: Fields,
-}
-
-/// Parsed reference of a schema.
-///
-/// As per specification a schema is referenced by its name and hash in a query.
-#[derive(Debug)]
-struct Schema {
-    /// Given name of this schema.
-    name: String,
-
-    /// Hash of the schema.
-    hash: Hash,
-}
-
-#[derive(Error, Debug)]
-#[allow(missing_copy_implementations)]
-pub enum ParseSchemaError {
-    #[error("Schema name is too long with {0} characters ({1} allowed)")]
-    TooLong(usize, usize),
-
-    #[error("Name and / or hash field is missing in string")]
-    MissingFields,
-
-    #[error(transparent)]
-    InvalidHash(#[from] p2panda_rs::hash::HashError),
-}
-
-impl Schema {
-    /// Parsing a schema string.
-    pub fn parse(str: &str) -> Result<Self, ParseSchemaError> {
-        // Allowed schema length is the maximum length of the schema name defined by the p2panda
-        // specification, the length of the separator and hash size times two because of its
-        // hexadecimal encoding.
-        let max_len = MAX_SCHEMA_NAME + SCHEMA_NAME_SEPARATOR.len() + (HASH_SIZE * 2);
-        if str.len() > max_len {
-            return Err(ParseSchemaError::TooLong(str.len(), max_len));
-        }
-
-        // Split the string by the separator and return the iterator in reversed form
-        let mut fields = str.rsplitn(2, SCHEMA_NAME_SEPARATOR);
-
-        // Since we start parsing the string from the back, we look at the hash first
-        let hash = fields
-            .next()
-            .ok_or(ParseSchemaError::MissingFields)?
-            .try_into()?;
-
-        // Finally parse the name of the given schema
-        let name = fields
-            .next()
-            .ok_or(ParseSchemaError::MissingFields)?
-            .to_string();
-
-        Ok(Self { name, hash })
-    }
-
-    /// Returns name of SQL table holding document views of this schema.
-    pub fn table_name(&self) -> String {
-        format!(
-            "{}{}{}",
-            self.name,
-            SCHEMA_NAME_SEPARATOR,
-            self.hash.as_str()
-        )
-    }
+    pub fields: Fields,
 }
 
 #[derive(Error, Debug)]
 #[allow(missing_copy_implementations)]
 pub enum AbstractQueryError {
     #[error(transparent)]
-    InvalidSchema(#[from] ParseSchemaError),
+    InvalidSchema(#[from] SchemaParseError),
 
     #[error("Query needs to contain a root operation definition")]
     OperationMissing,
@@ -174,18 +94,18 @@ pub enum AbstractQueryError {
 }
 
 #[derive(Debug)]
-struct AbstractQuery {
+pub struct AbstractQuery {
     /// Schema hash of the queried document view.
-    schema: Schema,
+    pub schema: Schema,
 
     /// Filter or sorting arguments which can be applied to the results.
-    arguments: Arguments,
+    pub arguments: Arguments,
 
     /// Fields which give general information on the given document view.
-    meta_fields: MetaFields,
+    pub meta_fields: MetaFields,
 
     /// Materialized field values of the document view.
-    fields: Fields,
+    pub fields: Fields,
 }
 
 impl AbstractQuery {
@@ -212,6 +132,7 @@ impl AbstractQuery {
                     let name = field
                         .name()
                         .ok_or(AbstractQueryError::ApplicationFieldInvalid)?
+                        .text()
                         .to_string();
 
                     match field.selection_set() {
@@ -268,10 +189,11 @@ impl AbstractQuery {
                     let name = field
                         .name()
                         .ok_or(AbstractQueryError::MetaFieldInvalid)?
+                        .text()
                         .to_string();
 
                     // No meta fields should have a selection set, except of "fields"
-                    if field.selection_set().is_some() && name.as_str() == "fields" {
+                    if field.selection_set().is_some() && name.as_str() != "fields" {
                         return Err(AbstractQueryError::MetaFieldInvalid);
                     }
 
@@ -314,6 +236,7 @@ impl AbstractQuery {
             let name = arg
                 .name()
                 .ok_or(AbstractQueryError::ArgumentInvalid)?
+                .text()
                 .to_string();
 
             // Expect arguments of known name and value type
@@ -428,135 +351,10 @@ impl AbstractQuery {
                     fields,
                 })
             } else {
-                return Err(AbstractQueryError::RootSelectionSetMissing);
+                Err(AbstractQueryError::RootSelectionSetMissing)
             }
         } else {
-            return Err(AbstractQueryError::OperationMissing);
+            Err(AbstractQueryError::OperationMissing)
         }
-    }
-}
-
-fn root_to_sql(root: AbstractQuery) -> Result<String> {
-    let mut query = Query::select();
-
-    if root.fields.is_some() {
-        for field in root.fields.unwrap() {
-            match field {
-                Field::Name(name) => {
-                    query.column(Alias::new(&name));
-                }
-                Field::Relation(_) => {
-                    panic!("Relations not supported yet");
-                }
-            }
-        }
-    }
-
-    if root.meta_fields.is_some() {
-        for meta in root.meta_fields.unwrap() {
-            match meta {
-                MetaField::DocumentHash => {
-                    query.column(Alias::new("document"));
-                }
-            }
-        }
-    }
-
-    query.from(Alias::new(&root.schema.table_name()));
-
-    if root.arguments.is_some() {
-        for arg in root.arguments.unwrap() {
-            match arg {
-                Argument::DocumentHash(value) => {
-                    query.and_where(Expr::col(Alias::new("document")).eq(value.as_str()));
-                }
-            }
-        }
-    }
-
-    let sql = query.to_string(sea_query::PostgresQueryBuilder);
-
-    Ok(sql)
-}
-
-pub fn gql_to_sql(query: &str) -> Result<String> {
-    // Parse GraphQL
-    let document = parse_graphql_query(query).unwrap();
-
-    // Convert GraphQL document to our own abstract query representation
-    let root = AbstractQuery::new_from_document(document)?;
-
-    // Convert to SQL query
-    let sql = root_to_sql(root).unwrap();
-
-    Ok(sql)
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{gql_to_sql, Schema};
-
-    #[test]
-    fn parse_schema() {
-        let schema = Schema::parse(
-            "festivalevents_0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b",
-        )
-        .unwrap();
-        assert_eq!(schema.name, "festivalevents");
-        assert_eq!(
-            schema.hash.as_str(),
-            "0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b"
-        );
-
-        let schema = Schema::parse(
-            "mul_tiple_separators_0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b",
-        )
-        .unwrap();
-        assert_eq!(schema.name, "mul_tiple_separators");
-        assert_eq!(
-            schema.hash.as_str(),
-            "0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b"
-        );
-    }
-
-    #[test]
-    fn invalid_schema() {
-        // Missing name
-        assert!(Schema::parse("_0020c6f23dbdc3b5c7b9ab46293111c48fc78b").is_err());
-
-        // Invalid hash (wrong encoding)
-        assert!(Schema::parse("test_thisisnotahash").is_err());
-
-        // Separator missing
-        assert!(Schema::parse(
-            "withoutseparator0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b"
-        )
-        .is_err());
-
-        // Invalid hash (too short)
-        assert!(Schema::parse("invalid_hash_0020c6f23dbdc3b5c7b9ab46293111c48fc78b").is_err());
-
-        // Schema name exceeds limit
-        assert!(Schema::parse("too_long_name_with_many_characters_breaking_the_64_characters_limit_0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b").is_err());
-    }
-
-    #[test]
-    fn parser() {
-        let query = "{
-            festival_events_0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b(document: \"0020b177ec1bf26dfb3b7010d473e6d44713b29b765b99c6e60ecbfae742de496543\") {
-                document
-                fields {
-                    title
-                    description
-                }
-            }
-        }";
-
-        let sql = gql_to_sql(query).unwrap();
-
-        assert_eq!(
-            sql,
-            "SELECT \"title\", \"description\", \"document\" FROM \"festival_events_0020c65567ae37efea293e34a9c7d13f8f2bf23dbdc3b5c7b9ab46293111c48fc78b\" WHERE \"document\" = '0020b177ec1bf26dfb3b7010d473e6d44713b29b765b99c6e60ecbfae742de496543'"
-        );
     }
 }
