+++ aquadoggo/src/graphql/mod.rs
@@ -1,9 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
 mod api;
+mod convert;
 mod query;
-mod sql;
 
 pub use api::{Endpoint, StaticSchema};
 pub use query::QueryRoot;
-pub use sql::gql_to_sql;
